added first map creation steps

+++ src/nxt/rbt/graph/Edge.java
@@ -1,12 +1,13 @@
 package nxt.rbt.graph;
 
 public class Edge  {
+	
 	private final String id; 
 	private final Vertex source;
 	private final Vertex destination;
-	private final int weight; 
+	private final float weight; 
 	
-	public Edge(String id, Vertex source, Vertex destination, int weight) {
+	public Edge(String id, Vertex source, Vertex destination, float weight) {
 		this.id = id;
 		this.source = source;
 		this.destination = destination;
@@ -23,7 +24,7 @@ public class Edge  {
 	public Vertex getSource() {
 		return source;
 	}
-	public int getWeight() {
+	public float getWeight() {
 		return weight;
 	}
